Use proper CMake mechanisms to set compiler settings
By setting CMAKE_CXX_STANDARD, CMake will figure out the right flag to
pass to the compiler being used as well as error if the compiler doesn't
support that standard.  The standard can also be changed on a per-target
basis.

Using add_compiler_options() is just a bit cleaner syntactically and
will apply the flags to C and C++.

Signed-off-by: Rick Altherr <kc8apf@kc8apf.net>
Signed-off-by: Tim 'mithro' Ansell <mithro@mithis.com>