xc7patch: Generate Xilinx BIT header in output bitstream
Xilinx tools generate and expect an additional header to be added to the
bitstream.  This header is the only difference between .bit and .bin
files and is not required by the part in any way.  OpenOCD only supports
reading .bit files so the easiest path to a using FOSS tools from FASM
down is to generate a .bit header in xc7patch's output. To distinguish
xc7patch-generated files from Xilinx-generated files, the source file
field includes a Generator tag indicating that the bitstream was produced by
xc7patch.  Vivado uses the same technique to record the Vivado version
in a Version tag.

Signed-off-by: Rick Altherr <kc8apf@kc8apf.net>