added scripts to generate the YAML description for xc7z family
The Tools do not include the possibility to create a debug
bitstream with the BITSTREAM.GENERAL.DEBUGBITSTREAM option when the
target FPGA belongs to the xc7z family.

The debug bitstream is needed to acquire all the information about the
configuration buses of the FPGA and get a YAML description of the device.
Therefore, we have modified the .tcl script to get the information about
the device by selecting all the configuration rows from which we
generate the YAML file.

The generate.py file is used to generate the YAML file starting from the
outputs of the .tcl script.

To run the xc7z version of the .tcl script the environmental variable
`XRAY_XC7Z` has to be set.

The gen_part_base_yaml has changed to return only the the idcode of the
device in case the xc7z part is chosen.

generate.tcl: added possibility to create YAML for xc7z family FPGAs

generate.py: script to generate YAML description file for xc7z family
FPGAs

Makefile.specimen: calls the YAML generator tool for xc7z family FPGAs

gen_part_base_yaml.cc: outputs only idcode when used for the xc7z FPGAs

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>