fuzzers: refactor fuzzer 51
The fuzzer is now todo files based. It is also using generic
generate.py script.
For now the fuzzer instantiates picorv and randluts besides
the logic under test. This is required for to deliver negative
examples to segmaker. Because of the size of the design this
fuzzer is taking really long to complete. We need to find a
smallest possible design that will be enough to generate the DB.

Signed-off-by: Karol Gugala <kgugala@antmicro.com>