README/Dependencies - Updated "Quickstart Guide"
Updated the Quickstart Guide section of the README as some deviations were
required when following this with a fresh Ubuntu 16.04 install.  Fixes to
''requirements.txt'' were required to remove the "futures" dependency which is
in the python3 standard library and is only intended for use with python2;
"fasm" now goes by "pyfasm".

Signed-off-by: Jake Mercer <jmercer04@qub.ac.uk>