bscan block throws errors in spartan7 cfg; fix by changing user reg param
When compiling the cfg fuzzer sub-test, it looks like perhaps the
spartan7 requires that the user confg parameter and the location
of the BSCANE2 block be correlated: that is,

BSCANE2 at X0Y0 => user register 1
BSCANE2 at X0Y1 => user register 2
...

and so forth. Assigning register 2 to the block at X0Y0 makes
vivado throw an error

This patch adds a workaround to incremente the block and user
register together.