063-gtp-common-conf: add IBUFDS_GTE2 primitive as well
This commit also solves an issue in the bitfilter function.
Given that the GTP_COMMON and GTP_COMMON_MID_[LEFT|RIGHT] have the same
bits, but in different frames, we need to add all of the possible frames
to the bitfilter callback.

This causes issues in determining the IN_USE feature, therefore, there
must be two different bitfilter callbacks, one for GTP_COMMON, and one
for GTP_COMMON_MID_[LEFT|RIGHT]

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>