fuzzers: 005: Makefile: Rename clean and clean_all comamnds
The 'make clean' command in the top-level Makefile calls 'clean' of all
fuzzers. The top-level Makefile does not source and settings file before and
therefore no XRAY_PART is set.

The 005's fuzzers only cleans data for a XRAY_PART file, which is not set.
Remove all fuzzer data during clean and only for a specific part during
clean_all.

Signed-off-by: Daniel Schultz <d.schultz@phytec.de>