settings: Handle return code correctly
"eval $(python3 script.py)" does not handle the return code correctly and would try to evaluate
the complete output of the Python traceback. Temporarily save the output in a variable and check
the return code. If the RC is unequal to zero, return the RC and leave the bash script.

Signed-off-by: Daniel Schultz <d.schultz@phytec.de>